chore: remove early abstractions to simplify project
