+++ Dockerfile
@@ -10,7 +10,7 @@ RUN go mod download
 COPY . .
 
 # Build the application
-RUN CGO_ENABLED=1 GOOS=linux go build -o /padmd_app ./cmd/app/main.go
+RUN CGO_ENABLED=1 GOOS=linux go build -o /padmd_app ./main.go
 
 EXPOSE 8080
