+++ go.mod
@@ -2,4 +2,4 @@ module github.com/dylanxhernandez/proto-padmd
 
 go 1.22.7
 
-require github.com/mattn/go-sqlite3 v1.14.24 // indirect
+require github.com/mattn/go-sqlite3 v1.14.24
