+++ templates/content.html
@@ -1,9 +1,4 @@
-{{define "title"}}Home{{end}}
-{{define "page-title"}}
-    HOME
-{{end}}
-{{define "page-content"}}
-    <br/> 
+{{ define "content" }}
     <div>
         <table class="striped">
     	<tbody>
@@ -23,4 +18,5 @@
     	</tbody>
         </table>
     </div>
-{{end}}
+{{ end }}
+
