+++ templates/form_submitted.html
@@ -1,0 +1,8 @@
+{{ define "content" }}
+    <div>
+        <div id="form-container">
+            Form Submitted!
+        </div>
+    </div>
+{{ end }}
+
